% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarePlot.R
\name{rarePlot}
\alias{rarePlot}
\title{Make a 3-panel plot with rarefaction curves from different samples. This is not currently intended for general usage and is tailored to the paper this package supports.}
\usage{
rarePlot(panA, panB, panC, plotname = "rarefaction.pdf", path = ".")
}
\arguments{
\item{panA}{An object from the \code{shannonRarefy()} function representing the rarefaction curve of the first sample.}

\item{panB}{An object from the \code{shannonRarefy()} function representing the rarefaction curve of the second sample.}

\item{panC}{An object from the \code{shannonRarefy()} function representing the rarefaction curve of the third sample.}

\item{plotname}{The filename given to the output plot. Defaults to \code{"rarefaction.pdf"}.}

\item{dir}{The location of the directory to which the output plot should be written. Defaults to current working directory.}
}
\value{
A three-panel plot, written to disk as a .pdf file
}
\description{
Make a 3-panel plot with rarefaction curves from different samples. This is not currently intended for general usage and is tailored to the paper this package supports.
}
