% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannonDiv.R
\name{shannonDiv}
\alias{shannonDiv}
\title{Calculate the Shannon-Weaver diversity index}
\usage{
shannonDiv(comm)
}
\arguments{
\item{comm}{A vector with the species ID for each individual in your sample. Each species is represented by a unique integer.}
}
\value{
A number representing the Shannon diversity of the input sample.
}
\description{
Calculate the Shannon-Weaver diversity index
}
