% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannonRarefy.R
\name{shannonRarefy}
\alias{shannonRarefy}
\title{Generate a rarefaction-type curve based on the Shannon-Weaver diversity index.}
\usage{
shannonRarefy(comm, nrep)
}
\arguments{
\item{comm}{A vector with the species ID for each individual in your sample. Each species is represented by a unique integer.}

\item{nrep}{A number representing the number of random resamplings, without replacement, of \code{comm} from which to generate the rarefaction curve.}
}
\value{
A dataframe with column \code{RAR} representing the rarefaction curve, column \code{SDS} being the standard deviation of the rarefaction curve, and column \code{HOB} representing the Shannon diversity of the full empirical dataset.
}
\description{
Generate a rarefaction-type curve based on the Shannon-Weaver diversity index.
}
